# Improved polygons_get_closest_point_info performance 