Changed the logic of the function polygons_get_closest_point_info in the NavMeshQueries3D class.

Instead of retriangulating each polygons, we iterate over each edge of the polygon, and check on which side lies the point we are looking for. If it's clockwise of all the sides it lies on the interior and we can just return the point. Otherwise, on the sides where it is on the exterior, we calculate the closest point for this edge and take the minimum of those.

All these timings are with the godot demo project [navigation_mesh_chunks](https://github.com/godotengine/godot-demo-projects/tree/master/2d/navigation_mesh_chunks), with different sizes of chunks.

**Timings**

All the timings below are been taken using [rdtsc](https://fr.wikipedia.org/wiki/RDTSC). The lower, the better.
All the averages are on at least 3500 iterations of the function.

Current implementation
Chunk Size 256: cycles Average : 44 000,      min: 26 000,      max: 114 500 
Chunk Size 128: cycles Average : 118 000,    min: 68 500,      max: 236 000
Chunk Size 32  : cycles Average : 1 340 000, min: 1 000 000, max: 2 750 000

New implementation
Chunk Size 256: cycles Average : 17 000,   min: 3 600,   max: 100 500   (x 2.5 speed)
Chunk Size 128: cycles Average : 21 000,   min: 4 500,   max:  94 600    (x 5.6 speed)
Chunk Size 32  : cycles Average : 128 000, min: 98 500, max: 426 000   (x 10.5 speed)

(The smaller the chunkSize, the more polygons on the navmesh)
The new implementation has two benefits : 
- less cycles per polygon which scales much better to huge navmeshes
- we early out if we find the point inside a polygon, which can reduce even further in some cases.